% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASM_Server.R
\name{ASM_check_workflow}
\alias{ASM_check_workflow}
\title{Checks Workflow Preload List Against Current App State}
\usage{
ASM_check_workflow(state, session, pll)
}
\arguments{
\item{state}{ASM state from \code{ASM_fetch_state()}}

\item{session}{Shiny session variable}

\item{pll}{Workflow preload list.}
}
\value{
A list with the following attributes 
following elements:
\itemize{
  \item{isgood:}      Boolean indicating the exit status of the function.
  \item{msgs:}        Messages to be passed back to the user.
  \item{chk_msgs:}    Results of check, this can contain \code{"tags()"}
  \item{deps_found:}  Boolean indicating if all resoures were found.
  \item{dep_table:}   Table of dependency information with the following columns:
  \itemize{
    \item{mod_ID:}      The formods module ID of the module that needs the resource.
    \item{dep_type :}   Type of resource dependency currently only \code{"ds"} for data source.
    \item{id       :}   The formods module ID supplying the resource.
    \item{res_label:}   The resource label.
    \item{res_found:}   Logical indicating if the resource was found or not.
  }
}
}
\description{
Compares the resource dependencies in a workflow preload list to
those currently avialble in the state to determine if the workflow can be run.
}
\examples{
# The ASM session won't have any components so the check below should fail
sess_res = ASM_test_mksession()
session = sess_res[["session"]]
state = FM_fetch_mod_state(id="ASM", session=session)

# The DW test merge should require resource labels that are not currently present
pll = formods::FM_read_yaml(system.file(package="formods", "preload", "workflow_DW_merge.yaml"))

cwf_res = ASM_check_workflow(state=state, session=session, pll=pll)

cwf_res
}
