% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASM_Server.R
\name{ASM_fetch_state}
\alias{ASM_fetch_state}
\title{Fetch State Manager State}
\usage{
ASM_fetch_state(id, input, session, yaml_file, yaml_section)
}
\arguments{
\item{id}{Shiny module ID}

\item{input}{Shiny input variable}

\item{session}{Shiny session variable}

\item{yaml_file}{cofiguration file}

\item{yaml_section}{Section of the yaml file with the module configuration}
}
\value{
list containing the current state of the app including default
values from the yaml file as well as any changes made by the user. The list
has the following structure:
\itemize{
\item{yaml:} Full contents of the supplied yaml file.
\item{MC:} Module components of the yaml file.
\item{ASM:}
\itemize{
  \item{isgood:} Boolean object indicating if the file was successfully loaded.
  \item{checksum:} This is an MD5 sum of the loaded state file
}
 \item{MOD_TYPE:} Character data containing the type of module \code{"ASM"}
 \item{id:} Character data containing the module id module in the session variable.
}
}
\description{
Merges default app options with the changes made in the UI
}
