% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASM_Server.R
\name{ASM_init_state}
\alias{ASM_init_state}
\title{Initialize ASM Module State}
\usage{
ASM_init_state(yaml_file, yaml_section, id)
}
\arguments{
\item{yaml_file}{App configuration file}

\item{yaml_section}{Section of the yaml file with the module configuration}

\item{id}{ID string for the module.}
}
\value{
list containing an empty ASM state
}
\description{
Creates a list of the initialized module state
}
