% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASM_Server.R
\name{ASM_mk_preload}
\alias{ASM_mk_preload}
\title{Make List of Current ASM State}
\usage{
ASM_mk_preload(state)
}
\arguments{
\item{state}{ASM state object}
}
\value{
list with the following elements
\itemize{
  \item{isgood:}       Boolean indicating the exit status of the function.
  \item{msgs:}         Messages to be passed back to the user.
  \item{yaml_list:}    Lists with preload components.
}
}
\description{
Converts the current ASM state into a preload list.
}
\examples{
sess_res = ASM_test_mksession()
state = sess_res$state
res = ASM_mk_preload(state)
}
