% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASM_Server.R
\name{ASM_onload}
\alias{ASM_onload}
\title{Updates ASM After State Load}
\usage{
ASM_onload(state, session)
}
\arguments{
\item{state}{ASM state object}

\item{session}{Shiny session variable}
}
\value{
List contianing the ASM state objects and shiny session object
}
\description{
Creates a list of the initialized module state
}
