% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASM_Server.R
\name{ASM_run_workflow}
\alias{ASM_run_workflow}
\title{Run Specified Workflow}
\usage{
ASM_run_workflow(state, session, wfl)
}
\arguments{
\item{state}{ASM state from \code{ASM_fetch_state()}}

\item{session}{Shiny session variable}

\item{wfl}{List contining details about the workflow}
}
\value{
The ASM state object with the results stored as a list in the field \code{rwf_res}
following elements:
\itemize{
  \item{isgood:}      Boolean indicating the exit status of the function.
  \item{msgs:}        Messages to be passed back to the user.
}
}
\description{
Called from download handler and used to write a saved state
value if that is null
}
\examples{
# Populating the session with data
ds_plf = c(system.file(package="formods", "preload", "ASM_preload_empty.yaml"),
           system.file(package="formods", "preload", "UD_preload.yaml"),
           system.file(package="formods", "preload", "DM_preload_empty.yaml"))
session = list()
res_ds = FM_app_preload(session=session, sources=ds_plf)

session = res_ds[["session"]]
state   = res_ds[["all_sess_res"]][["ASM"]][["state"]]

# Creating the workflow preload list
wf_pll = c(
  FM_read_yaml( system.file(package="formods", "preload", "ASM_preload.yaml")),
  FM_read_yaml( system.file(package="formods", "preload", "UD_preload.yaml")),
  FM_read_yaml( system.file(package="formods", "preload", "DM_preload_empty.yaml")),
  FM_read_yaml( system.file(package="formods", "preload", "DW_preload_empty.yaml")))

tmp_preload = tempfile(fileext=".yaml")
yaml::write_yaml(file=tmp_preload, x=wf_pll)

wfl = list(
  require_ds = TRUE,
  preload   = tmp_preload
)

# Running the workflow
state = ASM_run_workflow(state=state, session=session, wfl=wfl)
}
