% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASM_Server.R
\name{ASM_Server}
\alias{ASM_Server}
\title{Save State Server}
\usage{
ASM_Server(
  id,
  yaml_section = "ASM",
  yaml_file = system.file(package = "formods", "templates", "config.yaml"),
  react_state = NULL
)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the modules UI elements}

\item{yaml_section}{Section of the yaml file with the module configuration (\code{"UD"})}

\item{yaml_file}{Upload Data configuration file}

\item{react_state}{Variable passed to server to allow reaction outside of module (\code{NULL})}
}
\value{
UD Server object
}
\description{
Server function for the Save State Shiny Module
}
