% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASM_Server.R
\name{ASM_test_mksession}
\alias{ASM_test_mksession}
\title{Populate Session Data for Module Testing}
\usage{
ASM_test_mksession(session = list())
}
\arguments{
\item{session}{Shiny session variable (in app) or a list (outside of app)}
}
\value{
The ASM portion of the `all_sess_res` returned from \code{\link{FM_app_preload}}
}
\description{
Populates the supplied session variable for testing.
}
\examples{
sess_res = ASM_test_mksession()
}
\seealso{
\code{\link{FM_app_preload}}
}
