% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASM_Server.R
\name{ASM_test_workflow}
\alias{ASM_test_workflow}
\title{Test Specified Workflow}
\usage{
ASM_test_workflow(ds_plf, wf_plf)
}
\arguments{
\item{ds_plf}{Vector of preload yaml list files containing data loading}

\item{wf_plf}{Vector of preload yaml list files containing workflows}
}
\value{
List with the following elements:
\itemize{
  \item{isgood:}      Boolean indicating the exit status of the function.
  \item{msgs:}        Messages to be passed back to the user.
}
}
\description{
This will take a preload file for loading data and a separate preload file for a workflow and test whether it can be run.
}
\examples{
 ds_plf = c(system.file(package="formods", "preload", "ASM_preload.yaml"),
            system.file(package="formods", "preload", "UD_preload.yaml"),
            system.file(package="formods", "preload", "DM_preload_empty.yaml"))

 wf_plf = c(system.file(package="formods", "preload", "ASM_preload.yaml"),
            system.file(package="formods", "preload", "UD_preload.yaml"),
            system.file(package="formods", "preload", "DM_preload_empty.yaml"),
            system.file(package="formods", "preload", "DW_preload.yaml"))

 res = ASM_test_workflow(ds_plf=ds_plf, wf_plf=wf_plf)
}
