% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASM_Server.R
\name{ASM_write_state}
\alias{ASM_write_state}
\title{Write State to File for Saving}
\usage{
ASM_write_state(state, session, file)
}
\arguments{
\item{state}{ASM state from \code{ASM_fetch_state()}}

\item{session}{Shiny session variable}

\item{file}{File name to write zipped state.}
}
\description{
Called from download handler and used to write a saved state
value if that is null
}
