% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_Server.R
\name{DM_append_report}
\alias{DM_append_report}
\title{Append Report Elements}
\usage{
DM_append_report(state, rpt, rpttype, gen_code_only = FALSE)
}
\arguments{
\item{state}{DM state from \code{DM_fetch_state()}}

\item{rpt}{Report with the current content of the report which will be appended to in
this function. For details on the structure see the documentation for
\code{\link[onbrand]{template_details}}}

\item{rpttype}{Type of report to generate (supported "xlsx", "pptx", "docx").}

\item{gen_code_only}{Boolean value indicating that only code should be
generated (\code{FALSE}).}
}
\value{
list containing the following elements
\itemize{
 \item{isgood:}    Return status of the function.
 \item{hasrptele:} Boolean indicator if the module has any reportable elements.
 \item{code:}      Code to generate reporting elements.
 \item{msgs:}      Messages to be passed back to the user.
 \item{rpt:}       Report with any additions passed back to the user.
}
}
\description{
Appends report elements to a formods report.
}
\examples{
# We need a state object to use below
sess_res = DM_test_mksession()
state = sess_res$state

rpt = list(summary = list(), sheets=list())

rpt_res = DM_append_report(state,
  rpt     = rpt,
  rpttype = "xlsx")

# Shows if report elements are present
rpt_res$hasrptele

# Code chunk to generate report element
cat(paste(rpt_res$code, collapse="\n"))

# Tabular summary of data views
rpt_res$rpt$summary
}
\seealso{
\code{\link[formods:FM_generate_report]{formods::FM_generate_report()}}
}
