% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_Server.R
\name{DM_hasds}
\alias{DM_hasds}
\title{Check DM State For Datasets}
\usage{
DM_hasds(state)
}
\arguments{
\item{state}{DM state from \code{DM_fetch_state()}}
}
\value{
Logical TRUE if there is a dataset or FALSE otherwise.
}
\description{
Walks through the DM state object to see if there are any
datasets available
}
\examples{
sess_res = DM_test_mksession()
state = sess_res[["state"]]
DM_hasds(state)
}
