% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_Server.R
\name{DM_mk_preload}
\alias{DM_mk_preload}
\title{Make List of Current DM State}
\usage{
DM_mk_preload(state)
}
\arguments{
\item{state}{DM state object}
}
\value{
list with the following elements
\itemize{
  \item{isgood:}       Boolean indicating the exit status of the function.
  \item{msgs:}         Messages to be passed back to the user.
  \item{yaml_list:}    Lists with preload components.
}
}
\description{
Reads in the app state from yaml files.
}
\examples{
sess_res = DM_test_mksession()
state = sess_res$state
res = DM_mk_preload(state)
}
