% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_Server.R
\name{DM_onload}
\alias{DM_onload}
\title{Processes State After Loading}
\usage{
DM_onload(state, session)
}
\arguments{
\item{state}{DM state from \code{DM_fetch_state()}}

\item{session}{Shiny session variable}
}
\value{
List contianing the DM state object and shiny session object
}
\description{
When loading a saved analysis, this will process the state
object to account for differences that may apply between servers.
}
\examples{
sess_res = DM_test_mksession()
session = sess_res$session
state   = sess_res$state
state = DM_onload(state, session)
}
