% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_Server.R
\name{DM_test_mksession}
\alias{DM_test_mksession}
\title{Populate Session Data for Module Testing}
\usage{
DM_test_mksession(session = list())
}
\arguments{
\item{session}{Shiny session variable (in app) or a list (outside of app)}
}
\value{
The DM portion of the `all_sess_res` returned from \code{\link{FM_app_preload}}
}
\description{
Populates the supplied session variable for testing.
}
\examples{
session = shiny::MockShinySession$new()
sess_res = DM_test_mksession(session=session)
}
