% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{DW_append_report}
\alias{DW_append_report}
\title{Append Report Elements}
\usage{
DW_append_report(state, content, rpttype)
}
\arguments{
\item{state}{DW state from \code{DW_fetch_state()}}

\item{content}{Current content of the report which will be appended to in
this function. For details on the structer see the documentation for \code{\link{formods::FM_generate_report}}.}

\item{tpttype}{Type of report to generate (supported "xlsx").}
}
\value{
list containing the following elements
\itemize{
 \item{isgood:}    Return status of the function
 \item{hasrptele:} Boolean indicator if the module has any reportable elements
 \item{code:}      Data wrangling R command
 \item{msgs:}      Messages to be passed back to the user
 \item{content:}   Messages to be passed back to the user
}
}
\description{
Takes the current state of the app and appends data views to an
xlsx report object.
}
\seealso{
\code{\link{formods::FM_generate_report}}
}
