% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{DW_attach_ds}
\alias{DW_attach_ds}
\title{Attach Data Set to DW State}
\usage{
DW_attach_ds(state, DS)
}
\arguments{
\item{state}{DW state module.}

\item{DS}{Dataset list object with elements described by the DS field
returned by \code{UD_fetch_state()}.}
}
\value{
state with data set attached
}
\description{
Attaches a dataset to the DW state supplied.
}
