% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{dw_eval_element}
\alias{dw_eval_element}
\title{Evaluates Data Wrangling Generated Code}
\usage{
dw_eval_element(state, cmd)
}
\arguments{
\item{state}{DW state from \code{DW_fetch_state()}}

\item{cmd}{string containing the data wrangling command}
}
\value{
list with the following elements
\itemize{
 \item{isgood:} Return status of the function.
 \item{msgs:}   Messages to be passed back to the user.
 \item{DS:}     Wrangled dataset.
}
}
\description{
Takes the current state and a string containing a data
wranlging command and evaluates it.
}
