% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{DW_fetch_current_view}
\alias{DW_fetch_current_view}
\title{Fetches Current Data View}
\usage{
DW_fetch_current_view(state)
}
\arguments{
\item{state}{DW state from \code{DW_fetch_state()}}
}
\value{
List containing the details of the active data view. The structure
of this list is the same as the structure of \code{state$DW$views} in the output of
\code{DW_fetch_state()}.
}
\description{
Takes a DW state and returns the current active view
}
