% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{DW_mk_preload}
\alias{DW_mk_preload}
\title{Make List of Current DW State}
\usage{
DW_mk_preload(state)
}
\arguments{
\item{state}{DW state object}
}
\value{
list with the following elements
\itemize{
  \item{isgood:}       Boolean indicating the exit status of the function.
  \item{msgs:}         Messages to be passed back to the user.
  \item{yaml_list:}    Lists with preload components.
}
}
\description{
Converts the current DW state into a preload list.
}
\examples{
sess_res = DW_test_mksession()
state = sess_res$state
res = DW_mk_preload(state)
}
