% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{DW_new_view}
\alias{DW_new_view}
\title{New Data Wrangling View}
\usage{
DW_new_view(state)
}
\arguments{
\item{state}{DW state from \code{DW_fetch_state()}}
}
\value{
DW state object containing a new data view and that view set as the
current active view. See the help for \code{DW_fetch_state()} for view
format.
}
\description{
Appends a new empty data wrangling view to the DW state object
and makes this new view the active view.
}
