% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{DW_proc_pll}
\alias{DW_proc_pll}
\title{Adding Preload List  Component}
\usage{
DW_proc_pll(state, session = NULL, pll)
}
\arguments{
\item{state}{DW state from \code{DW_fetch_state()}}

\item{session}{Shiny session variable}

\item{pll}{Preload list either from a preload file or}
}
\value{
state with the preload list added. See the field
 \code{state[["DW"]][["res"]][["proc_pll"]]} for the exit status and any
 messages.
}
\description{
Adds the wrangling component specified in the preload list
}
