% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{DW_rebuild_current_view}
\alias{DW_rebuild_current_view}
\title{Forces Rebuild of Current Data View}
\usage{
DW_rebuild_current_view(state, session)
}
\arguments{
\item{state}{DW state object}

\item{session}{Shiny session variable}
}
\value{
State with current data view rebuilt. Any messages will be stored in
\code{state[["DW"]][["res"]][["rebuild_current_view"]][["msgs"]]}
}
\description{
Will attempt to rebuild data views when changes in the source
datasets have been detected.
}
\examples{
# JMH add examples
}
