% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{DW_rectify}
\alias{DW_rectify}
\title{Rebuilds Data Views After Source Changes}
\usage{
DW_rectify(state, session, view_ids = NULL)
}
\arguments{
\item{state}{DW state object}

\item{session}{Shiny session variable}

\item{view_ids}{List of view ids to mend or NULL to attempt to mend all}
}
\value{
State with dataviews rebuilt if dataset changes were detected
}
\description{
Will attempt to rebuild data views when changes in the source datasets have been detected
}
\examples{
sess_res = DW_test_mksession()
state    = sess_res$state
session  = sess_res$session
view_ids = names(state[["DW"]][["views"]])[1]
state = DW_rectify(
  state    = state, 
  session  = session, 
  view_ids = view_ids)
}
