% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{DW_Server}
\alias{DW_Server}
\title{Data Wrangling Server}
\usage{
DW_Server(
  id,
  id_ASM = "ASM",
  id_UD = "UD",
  FM_yaml_file = system.file(package = "formods", "templates", "formods.yaml"),
  MOD_yaml_file = system.file(package = "formods", "templates", "DW.yaml"),
  react_state = NULL
)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the modules UI elements}

\item{id_ASM}{ID string for the app state managment module used to save and load app states}

\item{id_UD}{ID string for the upload data module used to handle uploads or
the name of the list element in react_state where the data set is stored.}

\item{FM_yaml_file}{App configuration file with FM as main section.}

\item{MOD_yaml_file}{Module configuration file with DW as main section.}

\item{react_state}{Variable passed to server to allow reaction outside of
module (\code{NULL})}
}
\value{
DW Server object
}
\description{
Server function for the data wrangling module
}
\examples{
if(interactive()){
library(shiny)
library(formods)
library(shinydashboard)
#https://fontawesome.com/icons?from=io

ui <- dashboardPage(
  skin="red",
  dashboardHeader(title="Data Wrangling"),
  dashboardSidebar(
     sidebarMenu(
       menuItem("Data Wrangling",    tabName="main",  icon=icon("table"))
     )
  ),
  dashboardBody(
    tabItems(
       tabItem(tabName="main",
        fluidRow(
          htmlOutput(NS("DW",  "DW_ui_compact"))),
        fluidRow(tags$br()),
        fluidRow(
          verbatimTextOutput("ui_state"),width=12)
       )
     )
   )
 )

# Main app server
server <- function(input, output, session) {
  # Empty reactive object to track and react to
  # changes in the module state outside of the module
  react_FM = reactiveValues()

  # Creating upstream data for the UD module
  id_UD = "UD"
  res = UD_test_mksession(session, id_UD)
  react_FM[[id_UD]] = res[["rsc"]][[id_UD]]

  # Module server
  DW_Server(id="DW", id_UD = id_UD, react_state=react_FM)

  # Current state outside of the module
  output$ui_state  =  renderText({
    input$input_data_file
    uiele = paste(capture.output(str(react_FM[["DW"]])), collapse="\n")
  uiele})
}

shinyApp(ui, server)
}
}
