% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{DW_set_current_view}
\alias{DW_set_current_view}
\title{Sets Current Data View}
\usage{
DW_set_current_view(state, dw_view)
}
\arguments{
\item{state}{DW state from \code{DW_fetch_state()}}

\item{dw_view}{Data view list of the format returned from \code{DW_fetch_current_view()}
(see the structure of \code{state$DW$views} in the output of \code{DW_fetch_state()}).}
}
\value{
DW state object with the value of \code{dw_view} set to the current view id.
}
\description{
Takes a DW state and an updated view and sets that view to the
current view_id
}
