% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{DW_test_mksession}
\alias{DW_test_mksession}
\title{Populate Session Data for Module Testing}
\usage{
DW_test_mksession(session)
}
\arguments{
\item{session}{Shiny session variable (in app) or a list (outside of app)}
}
\value{
The DW portion of the `all_sess_res` returned from \code{\link{ASM_set_app_state}}
}
\description{
Populates the supplied session variable for testing.
}
\examples{
sess_res = DW_test_mksession(session=list())
}
\seealso{
\code{\link{ASM_set_app_state}}
}
