% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{dwrs_builder}
\alias{dwrs_builder}
\title{Builds a Data Wrangling R Statement From ui Elements:}
\usage{
dwrs_builder(state)
}
\arguments{
\item{state}{module state with all of the current ui elements populated}
}
\value{
list containing the following elements
\itemize{
 \item{isgood:} Return status of the function
 \item{cmd:}    Data wrangling R command
 \item{action:} The action being performed
 \item{desc:}   Verbose description of the action
 \item{msgs:}   Messages to be passed back to the user
}
}
\description{
Takes the current ui elements and constructs the appropriate
data wrangling command from the user input.
}
