% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{eval_element}
\alias{eval_element}
\title{Evaluates Data Wrangling Generated Code}
\usage{
eval_element(state, cmd)
}
\arguments{
\item{state}{data wrangling module state}

\item{cmd}{string containing the data wrangling command}
}
\value{
list containing an empty app state object
}
\description{
Creates a list of the initialized app state
}
