% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{fetch_hold}
\alias{fetch_hold}
\title{Fetches the Hold Status UI Element Supplied}
\usage{
fetch_hold(state, inputId = NULL)
}
\arguments{
\item{state}{module state with all of the current ui elements populated}

\item{inputId}{The input ID of the UI element that was put on hold}
}
\value{
Boolean value with the hold status
}
\description{
When some buttons are clicked they will change the state of the
system, but other UI components will not detect that change correctly. So those
triggers are put on hold. This will fetch hold status for a specified inputId
}
