% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{fetch_resource}
\alias{fetch_resource}
\title{Fetches Resource Object}
\usage{
fetch_resource(catalog = NULL, id = NULL, idx = NULL, res_label = NULL)
}
\arguments{
\item{catalog}{Catalog table returned from formods fetch functions (e.g. '\code{FM_fetch_ds()}).}

\item{id}{Module id.}

\item{idx}{Element id from module, can be set to \code{NULL} or \code{""} if the \code{res_label} option is set.}

\item{res_label}{Resource label, must be set to \code{NULL} or \code{""} to ignore the resource label and use \code{idx}.}
}
\value{
list containing the current dataset with the following format:
\itemize{
  \item{isgood:}  Boolean indicating the whether a resource object was found.
   \item{msgs:}   Any messages generated
  \item{res_row:} Row of catalog associated with resource or \code{NULL} if not found.
  \item{res_obj:} Name of resource object or \code{NULL} if not found.
}
}
\description{
Resources such as datasets can be specified using the module
ID and either identifier index or the resource label. This
function will attempt to rectify those and give the "best" answer. If a
resource label is supplied and exists uniquely for that module ID, that
object will be returned. If not then an error will be returned. If the
resource label is not specified and the index is specified and that exists
uniquely, that object will be returned. Otherwise an error will be returned.
}
\examples{
# We need a module state and a Shiny session variable
# to use this function:
sess_res = UD_test_mksession()
session = sess_res$session
state   = sess_res$state

DSV = FM_fetch_ds(state=state, session=session, ids=c("UD", "DM", "DW"), meta_only=TRUE)

fetch_resource(catalog=DSV[["catalog"]], id = "UD", idx="1", res_label = "")
}
