% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FG_Server.R
\name{FG_build}
\alias{FG_build}
\title{Evaluates Figure Generation Code}
\usage{
FG_build(
  state,
  del_row = NULL,
  cmd = NULL,
  element = "unknown",
  desc = "unknown"
)
}
\arguments{
\item{state}{FG state from \code{FG_fetch_state()}}

\item{del_row}{Row number to be deleted (NULL if no rows need to be deleted)}

\item{cmd}{String containing the plotting command.  Set to NULL to initialize a
new figure or force a rebuild after a dataset update.}

\item{element}{Short name for the figure element being performed, eg. point}

\item{desc}{Verbose description for the action being performed}
}
\value{
list with the following elements
\itemize{
 \item{isgood:} Return status of the function.
 \item{msgs:}   Messages to be passed back to the user.
 \item{pages:}  List with each element containing a ggplot object (\code{p}) and the code to generate that object (\code{code})
}
}
\description{
Takes the current state and rebuilds the active figure. If the
elements table has a row flagged for deletion, it will be deleted. If the
cmd input is not NULL it will attempt to append that element to the figure.
}
