% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FG_Server.R
\name{fg_eval}
\alias{fg_eval}
\title{Evaluates Figure Generation Generated Code}
\usage{
fg_eval(state, cmd)
}
\arguments{
\item{state}{FG state from \code{FG_fetch_state()}}

\item{cmd}{String containing the plotting command.  Set to NULL to initialize a
new figure or force a rebuild after a dataset update.}
}
\value{
list with the following elements
\itemize{
 \item{isgood:} Return status of the function.
 \item{msgs:}   Messages to be passed back to the user.
 \item{pages:}  List with each element containing a ggplot object (\code{p}) and the code to generate that object (\code{code})
}
}
\description{
Takes the current state and a string containing a data
wranlging command and evaluates it.
}
