% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FG_Server.R
\name{fg_eval_element}
\alias{fg_eval_element}
\title{Evaluates Figure Generation Generated Code}
\usage{
fg_eval_element(state, cmd)
}
\arguments{
\item{state}{data wrangling module state}

\item{cmd}{string containing the data wrangling command}
}
\value{
list with the following elements
\itemize{
 \item{isgood:} Return status of the function.
 \item{msgs:}   Messages to be passed back to the user.
 \item{p:}      plot 
}
}
\description{
Takes the current state and a string containing a data
wranlging command and evaluates it.
}
