% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FG_Server.R
\name{FG_fetch_current_fig}
\alias{FG_fetch_current_fig}
\title{Fetchs Current Figure}
\usage{
FG_fetch_current_fig(state)
}
\arguments{
\item{state}{FG state from \code{FG_fetch_state()}}
}
\value{
list containing the current figure
}
\description{
Takes an FG state and returns the ccurrent active figure
}
