% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FG_fetch_log_path}
\alias{FG_fetch_log_path}
\title{Fetches the path to the log file}
\usage{
FG_fetch_log_path(state)
}
\arguments{
\item{state}{module state after yaml read}
}
\value{
Character string with the path to the log file.
}
\description{
Use this to get the path to the formods log file
}
