% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FG_Server.R
\name{FG_fetch_state}
\alias{FG_fetch_state}
\title{Fetch Figure Generation State}
\usage{
FG_fetch_state(
  id,
  input,
  session,
  yaml_file,
  yaml_section,
  id_UD = NULL,
  id_DW = NULL,
  react_state
)
}
\arguments{
\item{id}{Shiny module ID}

\item{input}{Shiny input variable}

\item{session}{Shiny session variable}

\item{yaml_file}{cofiguration file}

\item{yaml_section}{Section of the yaml file with the module configuration}

\item{id_UD}{ID string for the upload data module used to handle uploads or the name of the list element in react_state where the data set is stored.}

\item{id_DW}{ID string for the data wrangling module to process any uploaded data}

\item{react_state}{Variable passed to server to allow reaction outside of module (`NULL`)}
}
\value{
list containing the current state of the app including default
values from the yaml file as well as any changes made by the user
}
\description{
Merges default app options with the changes made in the UI
}
