% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FG_Server.R
\name{FG_init_state}
\alias{FG_init_state}
\title{Initialize FG Module State}
\usage{
FG_init_state(FM_yaml_file, MOD_yaml_file, id, session)
}
\arguments{
\item{FM_yaml_file}{App configuration file with FM as main section.}

\item{MOD_yaml_file}{Module configuration file with MC as main section.}

\item{id}{Shiny module ID}

\item{session}{Shiny session variable}
}
\value{
list containing an empty app state object
}
\description{
Creates a list of the initialized module state
}
\examples{
# These would be the Shiny input and session variables
input   = list()
session = list()

state = FG_init_state(
   FM_yaml_file  = system.file(package = "formods",
                               "templates",
                               "formods.yaml"),
   MOD_yaml_file = system.file(package = "formods",
                               "templates",
                               "FG.yaml"),
   id              = "FG",
   session         = session)

state
}
