% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FG_le}
\alias{FG_le}
\title{Appends entry to log file}
\usage{
FG_le(state, entry)
}
\arguments{
\item{state}{module state after yaml read}

\item{entry}{text to add}
}
\description{
Add the supplied txt and the module type to the log file
}
