% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FG_Server.R
\name{FG_mk_preload}
\alias{FG_mk_preload}
\title{Make List of Current FG State}
\usage{
FG_mk_preload(state)
}
\arguments{
\item{state}{FG state object}
}
\value{
list with the following elements
\itemize{
  \item{isgood:}       Boolean indicating the exit status of the function.
  \item{msgs:}         Messages to be passed back to the user.
  \item{yaml_list:}    Lists with preload components.
}
}
\description{
Converts the current FG state into a preload list.
}
\examples{
sess_res = FG_test_mksession()
state = sess_res$state
res = FG_mk_preload(state)
}
