% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FG_Server.R
\name{FG_set_button_click_msgs}
\alias{FG_set_button_click_msgs}
\title{Sets State Message from Button Click}
\usage{
FG_set_button_click_msgs(state, msgs)
}
\arguments{
\item{state}{FG state from \code{FG_fetch_state()}}

\item{cmd}{Character vector of messages.}
}
\value{
state with button message set
}
\description{
Any errors that need to be passed back to the user can be set
with this function.
}
