% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FG_Server.R
\name{FG_set_current_fig}
\alias{FG_set_current_fig}
\title{Sets Current Figure}
\usage{
FG_set_current_fig(state, fig)
}
\arguments{
\item{state}{FG state from \code{FG_fetch_state()}}

\item{fig}{Figure list from \code{FG_fetch_current_fig}}
}
\value{
State with the current figure updated
}
\description{
Takes an FG state and a figure list and sets that figure list
as the value for the active figure
}
