% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FG_Server.R
\name{FG_test_mksession}
\alias{FG_test_mksession}
\title{Populate Session Data for Module Testing}
\usage{
FG_test_mksession(session = list())
}
\arguments{
\item{session}{Shiny session variable (in app) or a list (outside of app)}
}
\value{
The FG portion of the `all_sess_res` returned from \code{\link{ASM_set_app_state}}
}
\description{
Populates the supplied session variable for testing.
}
\examples{
sess_res = FG_test_mksession()
}
\seealso{
\code{\link{ASM_set_app_state}}
}
