% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FG_Server.R
\name{FG_update_checksum}
\alias{FG_update_checksum}
\title{Updates FG Module Checksum}
\usage{
FG_update_checksum(state)
}
\arguments{
\item{state}{FG state from \code{FG_fetch_state()}}
}
\value{
state with checksum updated.
}
\description{
Called after any changes to figures, this function will update
the checksum of the module. This allows other modules to determine if there
were any changes to the figures within it.
}
\examples{
# This will create a populated FG state object:
sess_res = FG_test_mksession(session=list())
state   = sess_res$state
state = FG_update_checksum(state)
}
