% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_build_comment}
\alias{FM_build_comment}
\title{Create RStudio Formatted Comments}
\usage{
FM_build_comment(level = 1, comment_str)
}
\arguments{
\item{level}{Integer (1 (default),2, or 3) indicating the section level of the comment.}

\item{comment_str}{Character object.}
}
\value{
Formatted comment.
}
\description{
Takes a character string and builds a comment so it will be
formatted as a section at the specified level in RStudio
}
\examples{
FM_build_comment(1, "This is a level 1 header")

FM_build_comment(2, paste0(rep("Long string repeated.", 5), collapse=" "))

}
