% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_fetch_app_code}
\alias{FM_fetch_app_code}
\title{Fetches the code to reproduce analysis}
\usage{
FM_fetch_app_code(session)
}
\arguments{
\item{session}{Shiny session variable}
}
\value{
list with the following elements:
\itemize{
  \item{isgood:} Boolean indicating the whether code generation was
  successful
  (\code{TRUE})
\item{msgs:} Any messages generated
\item{code:} Code to regenerate the app
}
}
\description{
Takes the current state of the app and builds a script to
reproduce the analysis within the app.
}
