% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_fetch_app_info}
\alias{FM_fetch_app_info}
\title{Fetches Informaiton About the App}
\usage{
FM_fetch_app_info(session)
}
\arguments{
\item{session}{Shiny session variable.}
}
\value{
List with information about the app with the following structure
\itemize{
  \item{uiele:} System information as UI elements to be used in shiny apps.
  \item{msgs:}  System information as text to be used in a report/terminal.
  \item{si_paclages} Dataframe with currently used packages.
}
}
\description{
Returns diagnostic information about the app
}
\examples{
# We need a Shiny session object to use this function:
id="UD"
sess_res = UD_test_mksession(session=list(), id=id)
session = sess_res$session
app_info  = FM_fetch_app_info(session)
app_info$msgs 
}
