% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_fetch_app_state}
\alias{FM_fetch_app_state}
\title{Fetches the App State}
\usage{
FM_fetch_app_state(session)
}
\arguments{
\item{session}{Shiny session variable.}
}
\value{
App state or NULL if it's not defined.
}
\description{
Returns the entire state of the App
}
\examples{
# We need a Shiny session object to use this function:
id="UD"
sess_res = UD_test_mksession(session=list(), id=id)
session = sess_res$session
app_state = FM_fetch_app_state(session)
app_state
}
