% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_fetch_current_mods}
\alias{FM_fetch_current_mods}
\title{Fetches Details About Current Modules}
\usage{
FM_fetch_current_mods()
}
\value{
list with details about the currently supported modules.
}
\description{
Use this to get information about the currently supported
modules. This includes short names, UI elements,
}
