% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_fetch_ds}
\alias{FM_fetch_ds}
\title{Fetches data sets from modules in the app}
\usage{
FM_fetch_ds(state, session, ids = NULL)
}
\arguments{
\item{state}{Current module state after yaml file has been read}

\item{session}{Shiny session variable}

\item{ids}{Vector of ID strings for the modules containing the datasets or
NULL for all datasets available.}
}
\value{
list containing the current dataset with the following format:
\itemize{
  \item{isgood:} Boolean indicating the whether a dataset was found
  (\code{FALSE})
  \item{ds:} List of datasets with element names corresponding to the
  R object name for that dataset. This has the following format
  \itemize{
    \item{label:}  Text label for the dataset (used to display to the user)
    \item{DS:}     Data frame with the dataset
    \item{DSMETA:} Data frame with metadata about the colunns of the
    dataset in \code{DS}.
    \item{code:} Code to generate the dataset.
    \item{checksum:} Module checksum when the dataset was pulled
    \item{DSchecksum:} Checksum of the dataframe in DS
  }
  \item{catalog:} Dataframe containing the a tabular catalog of the
  datasets found.
  \itemize{
    \item{label:} JMH
    \item{object:} JMH
    \item{MOD_TYPE:} JMH
    \item{id:} JMH
    \item{checksum:} Module checksum
    \item{DSchecksum:} Checksum of the dataset
    \item{code:} JMH
  }
  \item{modules:} List with an entry for each module. The element name is
  the short name. Each of these is a list with an entry that is the shiny module
  ID. For each of these there is a checksum. For example to access the
  checksum of a DW module with a  module ID of 'my_id', you would use the
  following: \code{res$DW$my_id}.
}
}
\description{
Loops through each specified module ID or all modules if no ID
was specified. For each ID, an attempt will be made to extract any datasets
available.
}
