% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_fetch_dsviews}
\alias{FM_fetch_dsviews}
\title{Fetchs data set views when both UD and DW modules are in use}
\usage{
FM_fetch_dsviews(state, id_UD, id_DW, react_state)
}
\arguments{
\item{state}{Current module state after yaml file has been read}

\item{id_UD}{ID string for the upload data module used to handle uploads or
the name of the list element in react_state where the data set is stored.}

\item{id_DW}{ID string for the data wrangling module used to process any uploaded data}

\item{react_state}{Variable passed to server to allow reaction outside of module (\code{NULL})}
}
\value{
list containing the current dataset with the following format:
\itemize{
  \item{isgood:} Boolean indicating the whether a dataset was found
  (\code{FALSE})
  \item{dsviews:}
    \itemize{
      \item{dsv_summary:} Dataframe summarizing the available views
      \item{contents:} List containing the contents of the data views
      \item{columns:} List containing the columns names from the contents.
      \item{checksum:} List containing a checksum of the data view. You can
      use this to detect changes in that view.
  }
\item{UD_checksum:} Checksum used to detect changes in the source dataset
from the UD module (\code{NULL} if not found).
\item{DW_checksum:} Checksum used to detect changes in any of the views
from the DW module (\code{NULL} if not found).
\item{active:} Data view key for the active data view. If no DW views were
round it will just be "default". If there are DW views it will return the
active DW view.
}
The named elements in the lists discussed in \code{dsviews}
are either "default" for the originally loaded dataset or a unique key
generated by the DW module.
}
\description{
The upload data and data wranling modules can be used together.
This will take the ids for both modules and return the different views
available. It will also return an "acive" view. This either the uploaded
dataset if no wrangling information is available or the last wrangled view.
}
