% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_fetch_mod_state}
\alias{FM_fetch_mod_state}
\title{Fetch the Module State}
\usage{
FM_fetch_mod_state(session, id)
}
\arguments{
\item{session}{Shiny session variable.}

\item{id}{ID string for the module.}
}
\value{
module state or NULL if it's not defined.
}
\description{
Fetches the module state from the userdata under the specified
id
}
