% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_fetch_user_files_path}
\alias{FM_fetch_user_files_path}
\title{Fetches the path to the user files}
\usage{
FM_fetch_user_files_path(state)
}
\arguments{
\item{state}{module state after yaml read}
}
\value{
Character string with the path to the log file.
}
\description{
Use this to get the path to the formods log file
}
