% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_find_DS}
\alias{FM_find_DS}
\title{Finds Dataset when both UD and DW modules are in use}
\usage{
FM_find_DS(id_UD, id_DW, react_state)
}
\arguments{
\item{id_UD}{ID string for the upload data module used to handle uploads or
the name of the list element in react_state where the data set is stored.}

\item{id_DW}{ID string for the data wrangling module to process any uploaded data}

\item{react_state}{Variable passed to server to allow reaction outside of module (`NULL`)}
}
\value{
list containing the current dataset with the following format:
values from the yaml file as well as any changes made by the user
\itemize{
  \item{isgood:} Boolean indicating the whether a dataset was found
  (`FALSE`)
  \item{dsm:} Module where the dataset was found or "Not Found"
  \item{contents:} Data frame containting the contents of the dataset
  (`NULL`)
  \item{columns:} Columns names from the contents or ('NULL')
  \item{object_name:} Object name for the dataset (e.g. DS)
  \item{checksum:} This is an MD5 sum of the contents element and can be
  used to detect changes in the loaded file.
}
}
\description{
The upload data and data wranling modules can be used together
but a user may opt to upload the data and skip the wrangling part. This will
take the ids for both modules and try to determine which one is being used.
The function first looks for a dataset in data wrangling module associated
with id_DW. If it finds a dataset there it returns that value. If it fails
it then looks for a dataset in the upload data module associated with id_UD.
If it fails ther it returns a list with the defaults below.
}
