% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_has_ds_changed}
\alias{FM_has_ds_changed}
\title{Compares DS Summary with Current State for Changes}
\usage{
FM_has_ds_changed(state, fdres, ids, react_state)
}
\arguments{
\item{state}{module state after yaml read}

\item{fdres}{Output of \code{FM_fetch_ds()}.}

\item{ids}{IDs the current module is}

\item{react_state}{Reactive shiny object (in app) or a list (outside of app) used to trigger reactions}
}
\value{
Logical indicating if a change has been
detected (\code{TRUE}) or not '(\code{FALSE})
}
\description{
Takes the result of \code{FM_fetch_ds()} and compares that to the
current react_state to determine if any datasets have changed for the given
module IDs.
}
\examples{
session = shiny::MockShinySession$new()
sess_res = FG_test_mksession(session=session)
state = sess_res$state
FM_has_ds_changed(state = state, fdres = NULL, ids = ("DW"), react_state=list())
}
