% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_le}
\alias{FM_le}
\title{Appends entry to log file}
\usage{
FM_le(state, entry, escape_braces = TRUE)
}
\arguments{
\item{state}{Module state after yaml read}

\item{entry}{Text to add}

\item{escape_braces}{Set to \code{TRUE} (default) to escape curly braces in the entry, set to \code{FALSE} to have the values interpreted.}
}
\description{
Add the supplied txt and the module type to the log file
}
\examples{
# We need a module state to use this function:
id="UD"
sess_res = UD_test_mksession(session=list(), id=id)
state   = sess_res$state
FM_le(state, "This is a message")
}
