% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_load_test_state}
\alias{FM_load_test_state}
\title{Populates Modles for Testing}
\usage{
FM_load_test_state(
  session,
  input,
  yaml_file = system.file(package = "formods", "templates", "config.yaml"),
  yaml_section_UD = "UD",
  yaml_section_DW = "DW",
  yaml_section_FG = "FG",
  id_ASM = "ASM",
  id_UD = "UD",
  id_DW = "DW",
  id_FG = "FG",
  react_state
)
}
\arguments{
\item{session}{Shiny session variable}

\item{input}{Shiny input variable}

\item{yaml_file}{App configuration file}

\item{id_UD}{ID string for the upload data module.}

\item{id_DW}{ID string for the data wrangling module.}

\item{id_FG}{ID string for the figure generation module.
the name of the list element in react_state where the data set is stored.}

\item{react_state}{Variable passed to server to allow reaction outside of}

\item{yaml_section}{Section of the yaml file with the module configuration.}
}
\value{
ggplot object
}
\description{
Loads modules and creates an app state for testing
}
