% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_mk_error_fig}
\alias{FM_mk_error_fig}
\title{Generates `ggplot` Object with Error Message}
\usage{
FM_mk_error_fig(msgs)
}
\arguments{
\item{msgs}{Vector of error messages}
}
\value{
ggplot object
}
\description{
Takes a vector of messages and returns a ggplot object with the
text in the figure. This can be used in automated figure generation to
cascade an error message to the end user.
}
