% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_notify}
\alias{FM_notify}
\title{Shiny Notification}
\usage{
FM_notify(state, session)
}
\arguments{
\item{state}{Module state generating the notification}

\item{session}{Shiny session variable}

\item{notify_id}{Unique string for this notification}

\item{notify_text}{Text to go in the notification}

\item{rpterrors}{Boolean variable to generate reports with errors.}

\item{timestamp}{Time when noficiation was created (Sys.time())}
}
\value{
Boolean variable indicating if the notification was triggered
}
\description{
Generates a notification that should only show once.
}
