% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_pause_screen}
\alias{FM_pause_screen}
\title{Starts Modal Screen Pause.}
\usage{
FM_pause_screen(state, session, message)
}
\arguments{
\item{state}{Current module state after yaml file has been read.}

\item{session}{Shiny session variable.}

\item{message}{Optional message for the pause.}
}
\description{
Starts Modal Screen Pause.
}
