% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_pretty_sort}
\alias{FM_pretty_sort}
\title{Centralized Sorting Function}
\usage{
FM_pretty_sort(unsrt_data)
}
\arguments{
\item{unsrt_data}{Unsorted data.}
}
\value{
sorted data
}
\description{
When displaying information in a pull down this function can
be used to sort those options.
}
\examples{
# This is the full path to a test data file:
data_file_local  =  system.file(package="formods", "test_data", "TEST_DATA.xlsx")
# Excel files need a sheet specification:
sheet           = "DATA"
# We will also attach the sheets along with it
df = readxl::read_excel(path=data_file_local, sheet=sheet)
# Regular sorting:
sort(unique(df$Cohort))
FM_pretty_sort(unique(df$Cohort))
}
