% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_proc_include}
\alias{FM_proc_include}
\title{Process Include Files in formods.yaml}
\usage{
FM_proc_include(state, session)
}
\arguments{
\item{state}{formods State object.}

\item{session}{Shiny session variable.}
}
\value{
No return value, sets include flag to indicate everythign has been
set.
}
\description{
Reads in the formods.yaml file and copies any include files
that are needed into the user directory.
}
\examples{
# We need a module state object to use this function:
sess_res = UD_test_mksession()
state = sess_res$state
session = sess_res$session
FM_proc_include(state, session)
}
