% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_read_yaml}
\alias{FM_read_yaml}
\title{Fetch Yaml Contents}
\usage{
FM_read_yaml(file)
}
\arguments{
\item{file}{yaml file.}
}
\value{
List containing the contents of the yaml file.
}
\description{
Wrapper for read_yaml() to make this function available to
modules dependent on formods.
}
\examples{
res = FM_read_yaml(
   system.file(package = "formods",
                         "templates",
                         "formods.yaml"))

res
}
