% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_reset_app}
\alias{FM_reset_app}
\title{Resets the App State}
\usage{
FM_reset_app(session)
}
\arguments{
\item{session}{Shiny session variable.}
}
\value{
session variable with app data removed.
}
\description{
Removes formods data from the app.
}
\examples{
# We need a Shiny session object to use this function:
sess_res = UD_test_mksession()
session = sess_res$session
app_state = FM_reset_app(session)
app_state
}
