% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_resume_screen}
\alias{FM_resume_screen}
\title{Starts Modal Screen Pause.}
\usage{
FM_resume_screen(state, session)
}
\arguments{
\item{state}{Current module state after yaml file has been read.}

\item{session}{Shiny session variable.}
}
\description{
Starts Modal Screen Pause.
}
