% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_set_app_state}
\alias{FM_set_app_state}
\title{Set the App State}
\usage{
FM_set_app_state(session, app_state, set_holds = TRUE)
}
\arguments{
\item{session}{Shiny session variable.}

\item{app_state}{Loaded app state.}

\item{set_holds}{If TRUE (default) the holds will be set for all of the
modules present in the app state.}
}
\value{
No return value, just updates the app state in the session variable.
}
\description{
Takes a loaded app state and overwrites the current app state
}
\examples{
# We need a Shiny session object to use this function:
id="UD"
sess_res = UD_test_mksession(session=list(), id=id)
session = sess_res$session
app_state = FM_fetch_app_state(session)
FM_set_app_state(session, app_state)
}
