% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_set_mod_state}
\alias{FM_set_mod_state}
\title{Set the Module State}
\usage{
FM_set_mod_state(session, id, state)
}
\arguments{
\item{session}{Shiny session variable}

\item{id}{ID string for the module.}

\item{state}{Module state to set.}
}
\description{
Sets the module state from the userdata under the specified
id
}
