% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_set_ui_msg}
\alias{FM_set_ui_msg}
\title{Sets Message in State from UI Processing}
\usage{
FM_set_ui_msg(state, msgs)
}
\arguments{
\item{state}{formods State object.}

\item{msgs}{Character vector of messages.}
}
\value{
state with ui message set.
}
\description{
Any errors that need to be passed back to the user can be set
with this function.
}
\examples{
# We need a module state object to use this function:
id="UD"
sess_res = UD_test_mksession(session=list(), id=id)
state = sess_res$state
state = FM_set_ui_msg(state, "Something happend.")
}
