% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_tc}
\alias{FM_tc}
\title{Rung Try/Catch and Process Results}
\usage{
FM_tc(cmd, tc_env, capture)
}
\arguments{
\item{cmd}{R command to evaluate in the try/catch block}

\item{tc_env}{list of with names corresponding to object names and
corresponding Values to define in the try/catch environment}

\item{capture}{Character vector of values to capture after the command is
successfully captured}
}
\description{
Add the supplied txt and the module type to the log file
}
