% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\docType{package}
\name{formods}
\alias{formods-package}
\alias{formods}
\title{formods: Shiny modules for common tasks.}
\description{
Shiny apps can often make use of the same key elements, this package
provides modules for common tasks (data upload, wrangling data, figure
generation and saving the app state). These modules can react and interact
as well as generate code to create reproducible analyses.
}
\seealso{
\url{https://formods.ubiquity.tools/}
}
\author{
\strong{Maintainer}: John Harrold \email{john.m.harrold@gmail.com} (\href{https://orcid.org/0000-0003-2052-4373}{ORCID})

}
