% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{formods_check}
\alias{formods_check}
\title{Checks `formods` Dependencies}
\usage{
formods_check(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical indicating if messages should be displayed}
}
\value{
List with the following elements:
\itemize{
  \item{all_found:}    Boolean indicating if all packages were found
  \item{found_pkgs:}   Character vector of found packages
  \item{missing_pkgs:} Character vector of missing packages
}
}
\description{
Looks at the suggested dependencies and checks to make sure
}
\examples{
fcres = formods_check()
}
