% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{has_changed}
\alias{has_changed}
\title{Detect if a UI element has changed}
\usage{
has_changed(ui_val = NULL, old_val = NULL, init_value = c(""))
}
\arguments{
\item{ui_val}{Current value from the UI.}

\item{old_val}{Last value of of the element.}

\item{init_value}{Default value for reading in UI data when it has not been
defined.}
}
\value{
Boolean result of the comparison
}
\description{
Takes a UI element value and an older value and determines if
it has been modified
}
