% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{icon_link}
\alias{icon_link}
\title{Creates Icon Link}
\usage{
icon_link(href, target = "_blank", icon_name = "circle-info")
}
\arguments{
\item{href}{URL to link to.}

\item{target}{New tab name.}

\item{icon_name}{Name of icon to use (arguemnt to shiny::icon, default: "circle-info")}
}
\value{
A list with a shiny.tag class that can be converted into an HTML string via as.character() and saved to a file with save_html()
}
\description{
Creates a link to a Shiny icon
}
\examples{
icon_link(href="https://formods.ubiquity.tools")
}
