% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_Server.R
\name{is_importable}
\alias{is_importable}
\title{Determines if a File is Importable}
\usage{
is_importable(file_name)
}
\arguments{
\item{file_name}{Name of file to check.}
}
\value{
Logical indicating if a file is an excel file or not.
}
\description{
Check to see if a file can be imported
}
\examples{
is_importable("file.xls")
is_importable("file.csv")
is_importable("file.bob")
}
