% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{is_installed}
\alias{is_installed}
\title{Determines if a Package is Installed}
\usage{
is_installed(pkgname)
}
\arguments{
\item{pkgname}{Name of package}
}
\value{
Logical indicating if the packages is installed or not
}
\description{
Determines if the specified package is installed.
}
\examples{
# This package should exist
is_installed('digest')

# This package should not exist
is_installed('bad package name')
}
