% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{is_shiny}
\alias{is_shiny}
\title{Determine if Object is Shiny Session Object}
\usage{
is_shiny(session = list())
}
\arguments{
\item{session}{Session object}
}
\value{
Logical indicating if the object is a shiny session object or not
}
\description{
Determines if the specified object is a shiny session object.
}
\examples{
is_shiny(session = list())
}
