% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCA_UI.R
\name{NCA_UI}
\alias{NCA_UI}
\title{NCA UI}
\usage{
NCA_UI(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function}
}
\value{
return
}
\description{
UI function for the NCA Shiny Module
}
