% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{new_module_template}
\alias{new_module_template}
\title{Makes Template Files for formods New Module}
\usage{
new_module_template(
  SN = "NM",
  Module_Name = "New Module",
  package = "pkgname",
  element = "analysis",
  file_dir = tempdir()
)
}
\arguments{
\item{Module_Name}{Module long name}

\item{package}{Name of package that will contain the module}

\item{element}{What you would call the thing the module provides for example
the FG module provides "figures", the DW module provides "data views".}

\item{file_dir}{Directory to save file}

\item{MN}{Module short name}
}
\value{
list with the following elements:
\itemize{
\item{mc:}     Model components.
\item{server:} Server.R file.
\item{yaml:}   Yaml configureation file.
}
Each of these is a list with paths to the respective files:
\itemize{
\item{source:}     Template source.
\item{dest:}       Destination file name.
\item{dest_full:}  Full path to the destination file name.
}
}
\description{
If you want to create a new formods module this function will
create the template files for you.
}
\examples{
new_module_template()

}
