% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{render_str}
\alias{render_str}
\title{Evaluate R Code in String}
\usage{
render_str(estr = "")
}
\arguments{
\item{estr}{String to render.}
}
\value{
String containing the evaled as a character or the original string
}
\description{
Attempts to evaluate a string as a chunk of R code. If that
succeeds it will return the result. If not it will return the original text.
}
\examples{
res = render_str(estr="ls()")
}
