% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{run_formods}
\alias{run_formods}
\title{Run the {formods} Shiny App}
\usage{
run_formods(
  host = "127.0.0.1",
  port = 3838,
  server_opts = list(shiny.maxRequestSize = 30 * 1024^2),
  devmode = FALSE,
  mksession = FALSE
)
}
\arguments{
\item{host}{Hostname of the server ("127.0.0.1")}

\item{port}{Port number for the app (3838)}

\item{server_opts}{List of options (names) and their vlues (value) e.g.
\code{list(shiny.maxRequestSize = 30 * 1024^2)}.}

\item{devmode}{Boolean value, when TRUE will run formods with development
modules.}

\item{mksession}{Boolean value, when TRUE will load test session data
for app testing.}
}
\value{
Nothing is returned, this function just runs the built-in formods 
app.
}
\description{
Runs the test formods. app.
}
\examples{
if (interactive()) {
run_formods()
}
}
