% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{set_hold}
\alias{set_hold}
\title{Sets Hold on One or All UI Elements}
\usage{
set_hold(state, inputId = NULL)
}
\arguments{
\item{state}{module state with all of the current ui elements populated}

\item{inputId}{The input ID of the UI element that was put on hold or
\code{NULL} to hold all IDs in the module}
}
\value{
state with hold or holds set
}
\description{
When some buttons are clicked they will change the state of the
system, but other UI components will not detect that change correctly. So those
triggers are put on hold. This will set the hold for a specified inputId or
all ids if that value is set to NULL
}
