% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{set_wds}
\alias{set_wds}
\title{Sets/Updates Wrangled Dataset Field (WDS)}
\usage{
set_wds(state, WDS)
}
\arguments{
\item{state}{module state with all of the current ui elements populated}

\item{WDS}{new wrangled data set}
}
\value{
object with factors removed
}
\description{
Takes an object that is a factor and returns an unfactored
vector with the same type by the value removed
}
