% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UD_Server.R
\name{UD_attach_ds}
\alias{UD_attach_ds}
\title{Attach Data Set to UD State}
\usage{
UD_attach_ds(
  state,
  isgood = TRUE,
  load_msg = NULL,
  data_file_local = NULL,
  data_file_ext = NULL,
  data_file = NULL,
  sheet = NULL,
  sheets = NULL,
  code = "",
  object_name = NULL,
  contents = NULL
)
}
\arguments{
\item{state}{UD state module.}

\item{isgood}{Boolean object indicating if the file was successfully loaded.}

\item{load_msg}{Text message indicated the success or any problems encountered when uploading the file.}

\item{data_file_local}{Full path to the data file on the server.}

\item{data_file_ext}{File extension of the uploaded file.}

\item{data_file}{Dataset file name without the path.}

\item{sheet}{If the uploaded file is an excel file, this is the currently selected sheet.}

\item{sheets}{If the uploaded file is an excel file, this is a character vector of the sheets present in that file.}

\item{code}{Code to load dataset.}

\item{object_name}{Name of the dataset object created when code is evaluated.}

\item{contents}{Data frame containting the contents of the data file.}
}
\value{
state with data set attached
}
\description{
Attaches a dataset to the UD state supplied.
}
