% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UD_Server.R
\name{UD_fetch_code}
\alias{UD_fetch_code}
\title{Fetch Module Code}
\usage{
UD_fetch_code(state)
}
\arguments{
\item{state}{UD state from \code{UD_fetch_state()}}
}
\value{
Character object vector with the lines of code
}
\description{
Fetches the code to generate results seen in the app
}
\examples{
# This creates a session variable that will be available in Shiny
state = UD_test_mksession(session=list())$state
UD_fetch_code(state)
}
