% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UD_Server.R
\name{UD_fetch_state}
\alias{UD_fetch_state}
\title{Fetch Upload Data State}
\usage{
UD_fetch_state(id, input, session, yaml_file, yaml_section)
}
\arguments{
\item{id}{Shiny module ID}

\item{input}{Shiny input variable}

\item{session}{Shiny session variable}

\item{yaml_file}{cofiguration file}

\item{yaml_section}{Section of the yaml file with the module configuration}
}
\value{
list containing the current state of the app including default
values from the yaml file as well as any changes made by the user
}
\description{
Merges default app options with the changes made in the UI
}
