% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UD_Server.R
\name{UD_init_state}
\alias{UD_init_state}
\title{Initialize UD Module State}
\usage{
UD_init_state(FM_yaml_file, MOD_yaml_file, id)
}
\arguments{
\item{FM_yaml_file}{App configuration file with FM as main section.}

\item{MOD_yaml_file}{Module configuration file with MC as main section.}

\item{id}{ID string for the module.}
}
\value{
list containing an empty UD state
}
\description{
Creates a list of the initialized module state
}
