% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{unfactor}
\alias{unfactor}
\title{Remove Factor From Object}
\usage{
unfactor(fctobj)
}
\arguments{
\item{fctobj}{Factorized object}
}
\value{
object with factors removed
}
\description{
Takes an object that is a factor and returns an unfactored
vector with the same type by the value removed
}
