% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{was_clicked}
\alias{was_clicked}
\title{Detect if a UI Button Has Been Clicked}
\usage{
was_clicked(ui_value = NULL, counter = NULL, init_value = c(""))
}
\arguments{
\item{ui_value}{Current value from the button UI.}

\item{counter}{Current value of the button counter.}

\item{init_value}{Default value for reading in UI data when it has not been
defined.}
}
\value{
object with factors removed
}
\description{
Takes a UI element value and a button counter and determines if
it has been clicked
}
