% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASM_Server.R
\name{ASM_fetch_dlfn}
\alias{ASM_fetch_dlfn}
\title{Fetch Download File Name}
\usage{
ASM_fetch_dlfn(state, extension = ".zip")
}
\arguments{
\item{state}{ASM state from \code{ASM_fetch_state()}}

\item{extension}{File extension for the download (default: ".zip")}
}
\value{
character object with the download file name
}
\description{
Gets either the file name specified by the user or the default
value if that is null
}
\examples{
# Creating a state object for testing
sess_res = ASM_test_mksession(session=list(), full_session=FALSE)
state = sess_res$state
dlfn = ASM_fetch_dlfn(state)
dlfn
}
