% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{DW_add_wrangling_element}
\alias{DW_add_wrangling_element}
\title{Adding Wrangling Element to Current Data View}
\usage{
DW_add_wrangling_element(state, dwb_res, dwee_res)
}
\arguments{
\item{state}{DW state from \code{DW_fetch_state()}}

\item{dwb_res}{Output from \code{dwrs_builder()}}

\item{dwee_res}{Output from \code{dw_eval_element()}
returned by \code{UD_fetch_state()}.}
}
\value{
state with data set attached
}
\description{
Adds the wrangling element to the current data view.
}
