% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{DW_attach_ds}
\alias{DW_attach_ds}
\title{Attach Data Set to DW State}
\usage{
DW_attach_ds(state, id_UD, session)
}
\arguments{
\item{state}{DW state from \code{DW_fetch_state()}}

\item{id_UD}{ID string for the upload data module used to handle uploads or
returned by \code{UD_fetch_state()}.}

\item{id}{Shiny module ID}
}
\value{
state with data set attached
}
\description{
Attaches a dataset to the DW state supplied.
}
