% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{DW_attach_ds}
\alias{DW_attach_ds}
\title{Attach Data Set to DW View}
\usage{
DW_attach_ds(state, dw_view, ds_source_id)
}
\arguments{
\item{state}{DW state from \code{DW_fetch_state()}}

\item{dw_view}{Data view list of the format returned from \code{DW_fetch_current_view()}}

\item{ds_source_id}{Source id from \code{DSV[["catalog"]][["object"]]} for
the dataset.}
}
\value{
state with data set attached
}
\description{
Attaches a dataset to the DW state supplied.
}
\examples{
# Within shiny both session and input variables will exist,
# this creates examples here for testing purposes:
sess_res = DW_test_mksession()
session = sess_res$session
input   = sess_res$input

# We also need a state variable
state = sess_res$state

# We need to identify the UD module with the data
# JMH TODO: update example
dw_view = DW_fetch_current_view(state=state)

ds_source_id = names(state[["DW"]][["DSV"]][["ds"]])[1]

dw_view = DW_attach_ds(state=state, dw_view=dw_view, ds_source_id=ds_source_id)
}
