% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{DW_fetch_code}
\alias{DW_fetch_code}
\title{Fetch Module Code}
\usage{
DW_fetch_code(state)
}
\arguments{
\item{state}{DW state from \code{DW_fetch_state()}}
}
\value{
Character object vector with the lines of code
and isgood)
}
\description{
Fetches the code to generate results seen in the app
}
\examples{
# This will create a formods DW state object for the example
sess_res = DW_test_mksession()
state   = sess_res$state
code = DW_fetch_code(state)
cat(code)
}
