% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{DW_init_state}
\alias{DW_init_state}
\title{Initialize App State}
\usage{
DW_init_state(yaml_file, yaml_section)
}
\arguments{
\item{yaml_file}{App cofiguration file}

\item{yaml_section}{Section of the yaml file with the module configuration}
}
\value{
list containing an empty app state object
}
\description{
Creates a list of the initialized app state
}
