% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{DW_init_state}
\alias{DW_init_state}
\title{Initialize DW Module State}
\usage{
DW_init_state(yaml_file, yaml_section, id_UD, react_state)
}
\arguments{
\item{yaml_file}{App configuration file}

\item{yaml_section}{Section of the yaml file with the module configuration}

\item{id_UD}{ID string for the upload data module used to handle uploads or
the name of the list element in react_state where the data set is stored.}

\item{react_state}{Variable passed to server to allow reaction outside of
module (\code{NULL})}
}
\value{
list containing an empty DW state
}
\description{
Creates a list of the initialized module state
}
