% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{DW_new_view}
\alias{DW_new_view}
\title{New Data Wrangling View}
\usage{
DW_new_view(state, id_UD, react_state)
}
\arguments{
\item{state}{DW state from \code{DW_fetch_state()}}

\item{id_UD}{ID string for the upload data module used to handle uploads or
the name of the list element in react_state where the data set is stored.}

\item{react_state}{Variable passed to server to allow reaction outside of module (\code{NULL})}
}
\value{
DW state object containing a new data view and that view set as the
current active view. See the help for \code{DW_fetch_state()} for view
format.
}
\description{
Appends a new empty data wrangling view to the DW state object
and makes this new view the active view.
}
