% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{DW_Server}
\alias{DW_Server}
\title{Data Wrangling Server}
\usage{
DW_Server(
  id,
  yaml_section = "DW",
  yaml_file = system.file(package = "formods", "templates", "config.yaml"),
  id_UD = "UD",
  react_state = NULL
)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the modules UI elements}

\item{yaml_section}{Section of the yaml file with the module
configuration (\code{"DW"})}

\item{yaml_file}{Upload Data cofiguration file}

\item{id_UD}{ID string for the upload data module used to handle uploads or
the name of the list element in react_state where the data set is stored.}

\item{react_state}{Variable passed to server to allow reaction outside of
module (\code{NULL})}
}
\value{
DW Server object
}
\description{
Server function for the data wrangling module
}
