% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_Server.R
\name{dwrs_builder}
\alias{dwrs_builder}
\title{Builds a Data Wrangling R Statement From ui Elements:}
\usage{
dwrs_builder(state)
}
\arguments{
\item{state}{module state with all of the current ui elements populated}
}
\value{
list containing an empty app state object
}
\description{
Creates a list of the initialized app state
}
