% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FG_Server.R
\name{FG_extract_page}
\alias{FG_extract_page}
\title{Extracts Specific Page from Paginated Figure}
\usage{
FG_extract_page(state, page)
}
\arguments{
\item{state}{FG state from \code{FG_fetch_state()}}

\item{page}{Page number to extract}
}
\value{
state with checksum updated.
}
\description{
Used to extract the specified page from the current figure.
}
