% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FG_Server.R
\name{FG_Server}
\alias{FG_Server}
\title{Figure Generation Server}
\usage{
FG_Server(
  id,
  yaml_section = "FG",
  yaml_file = system.file(package = "formods", "templates", "config.yaml"),
  id_ASM = "ASM",
  id_UD = "UD",
  id_DW = "DW",
  react_state = NULL
)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function}

\item{yaml_section}{Section of the yaml file with the module configuration (\code{"FG"})}

\item{yaml_file}{Upload Data cofiguration file}

\item{id_ASM}{ID string for the app state management module used to save and load app states}

\item{id_UD}{ID string for the upload data module used to handle uploads or the name of the list element in react_state where the data set is stored.}

\item{id_DW}{ID string for the data wrangling module to process any uploaded data}

\item{react_state}{Variable passed to server to allow reaction outside of module (\code{NULL})}
}
\value{
FG Server object
}
\description{
Server function for the figure generation module
}
