% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_fetch_log_path}
\alias{FM_fetch_log_path}
\title{Fetches the path to the log file}
\usage{
FM_fetch_log_path(state)
}
\arguments{
\item{state}{module state after yaml read}
}
\value{
Character string with the path to the log file.
}
\description{
Use this to get the path to the formods log file
}
\examples{
# This function assumes that some module state exists:
state = UD_init_state(
 FM_yaml_file  = system.file(package = "formods",
                             "templates",
                             "formods.yaml"),
 MOD_yaml_file = system.file(package = "formods",
                             "templates",
                             "UD.yaml"),
 id = "UD")
FM_fetch_log_path(state)
}
