% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_fetch_log_path}
\alias{FM_fetch_log_path}
\title{Fetches the Path to the Log File}
\usage{
FM_fetch_log_path(state)
}
\arguments{
\item{state}{module state after yaml read}
}
\value{
Character string with the path to the log file.
}
\description{
Use this to get the path to the formods log file
}
\examples{
# Within shiny a session variable will exist,
# this creates one here for testing purposes:
sess_res = UD_test_mksession()
session = sess_res$session
state   = sess_res$state
FM_fetch_log_path(state)
}
