% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_init_state}
\alias{FM_init_state}
\title{Initialize a formods State Object}
\usage{
FM_init_state(
  FM_yaml_file,
  MOD_yaml_file,
  id,
  MT,
  button_counters,
  ui_ids,
  ui_hold
)
}
\arguments{
\item{FM_yaml_file}{App configuration file with FM as main section.}

\item{MOD_yaml_file}{Module configuration file with MC as main section.}

\item{id}{Shiny module ID.}

\item{MT}{Type of module using the short name (e.g. "UD", "FG", etc.).}

\item{button_counters}{Vector of button UI elements that need to be tracked.}

\item{ui_ids}{List of UI ids in the model.}

\item{ui_hold}{Vector of UI elements that require holding.}
}
\description{
Initializes a formods state object with common elements.
}
