% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_mk_app_preload}
\alias{FM_mk_app_preload}
\title{Preload Data Into App}
\usage{
FM_mk_app_preload(session)
}
\arguments{
\item{session}{Shiny session variable (in app) or a list (outside of app)}
}
\value{
list with the following elements
\itemize{
  \item{isgood:}       Boolean indicating the exit status of the function.
  \item{msgs:}         Messages to be passed back to the user.
  \item{yaml_list:}    Lists with preload components.
  \item{session:} Returning the session variable to be used in scripting (not in app).
}
}
\description{
Populates session data for testing or to load a specific
analysis.
}
\examples{
sources=system.file(package="formods", "preload", "UD.yaml")
sess_res = ASM_set_app_state(session=list(), sources=sources)
mkp_res = ASM_mk_app_preload(sess_res$session)
}
