% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_pause_screen}
\alias{FM_pause_screen}
\title{Starts Modal Screen Pause}
\usage{
FM_pause_screen(state, session, message)
}
\arguments{
\item{state}{Current module state after yaml file has been read.}

\item{session}{Shiny session variable.}

\item{message}{Optional message for the pause.}
}
\value{
Pauses the screen and has no return value.
}
\description{
Start a modal screen pause.
}
\examples{
# We need a module state object and Shiny session objects to use this function:
sess_res = UD_test_mksession(session=list())
session = sess_res$session
state = sess_res$state
FM_pause_screen(state, session)
FM_resume_screen(state, session)
}
