% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_proc_include}
\alias{FM_proc_include}
\title{Sets Message in State from UI Processing}
\usage{
FM_proc_include(state, session)
}
\arguments{
\item{state}{formods State object.}

\item{session}{Shiny session variable.}
}
\value{
No return value, sets message in supplied session variable.
}
\description{
Any errors that need to be passed back to the user can be set
with this function.
}
\examples{
# We need a module state object to use this function:
id="UD"
sess_res = UD_test_mksession(session=list(), id=id)
state = sess_res$state
session = sess_res$session
FM_proc_include(state, session)
}
