% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_resume_screen}
\alias{FM_resume_screen}
\title{Stops Modal Screen Pause}
\usage{
FM_resume_screen(state, session)
}
\arguments{
\item{state}{Current module state after yaml file has been read.}

\item{session}{Shiny session variable.}
}
\value{
No return value, called to disable screen pause.
}
\description{
Stops Modal Screen Pause
}
\examples{
# We need a module state object and Shiny session objects to use this function:
sess_res = UD_test_mksession(session=list())
session = sess_res$session
state = sess_res$state
FM_pause_screen(state, session)
FM_resume_screen(state, session)
}
