% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_set_app_state}
\alias{FM_set_app_state}
\title{Set the App State}
\usage{
FM_set_app_state(session, app_state, set_holds = TRUE)
}
\arguments{
\item{session}{Shiny session variable.}

\item{app_state}{Loaded app state.}

\item{set_holds}{If TRUE (default) the holds will be set for all of the
modules present in the app state.}
}
\description{
Takes a loaded app state and overwrites the current app state
}
