% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_set_notification}
\alias{FM_set_notification}
\title{Shiny Notification}
\usage{
FM_set_notification(state, notify_text, notify_id, type = "info")
}
\arguments{
\item{state}{Module state generating the notification}

\item{notify_text}{Text to go in the notification}

\item{notify_id}{Unique string for this notification}

\item{type}{- Can be either "success", "failure", "info" (default), or
"warning"}
}
\value{
Module state with notification text set
}
\description{
Generates a notification that should only show once.
}
