% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formods.R
\name{FM_set_ui_msg}
\alias{FM_set_ui_msg}
\title{Sets Message in State from UI Processing}
\usage{
FM_set_ui_msg(state, msgs, append = FALSE)
}
\arguments{
\item{state}{formods State object.}

\item{msgs}{Character vector of messages.}

\item{append}{When \code{TRUE}, msgs will be appended to any current messages. When
\code{FALSE} (default) msgs will replace any existing messaages.}
}
\value{
state with ui message set.
}
\description{
Any errors that need to be passed back to the user can be set
with this function.
}
\examples{
# We need a module state object to use this function:
sess_res = UD_test_mksession()
state = sess_res$state
state = FM_set_ui_msg(state, "Something happend.")
}
