% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UD_Server.R
\name{UD_ds_read}
\alias{UD_ds_read}
\title{Generate Code and Load DS}
\usage{
UD_ds_read(
  state,
  data_file_ext = NULL,
  data_file_local = NULL,
  data_file = NULL,
  sheets = NULL,
  sheet = NULL
)
}
\arguments{
\item{data_file_ext}{File extension of the uploaded file.}

\item{data_file_local}{Full path to the data file on the server.}

\item{data_file}{Dataset file name without the path.}

\item{sheets}{If the uploaded file is an excel file, this is a character vector of the sheets present in that file.}

\item{sheet}{If the uploaded file is an excel file, this is the currently selected sheet.}
}
\value{
list with the elements of the dataset (contents, object_name, code,
and isgood)
}
\description{
Generates the code for loading a dataset and returns both the
code and the contents
}
