% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UD_Server.R
\name{UD_fetch_ds}
\alias{UD_fetch_ds}
\title{Fetch Module Datasets}
\usage{
UD_fetch_ds(state)
}
\arguments{
\item{state}{UD state from \code{UD_fetch_state()}}
}
\value{
Character object vector with the lines of code

list containing the following elements
\itemize{
 \item{isgood:}    Return status of the function.
 \item{hasds:}     Boolean indicator if the module has any datasets
 \item{msgs:}      Messages to be passed back to the user.
 \item{ds:}        List with datasets. Each list element has the name of
 the R-object for that dataset. Each element has the following structure:
 \itemize{
   \item{label:}
   \item{MOD_TYPE:}
   \item{id:}
   \item{DSMETA:}
   \item{code:}
   \item{checksum:}
   \item{DSchecksum:}
 }
}
}
\description{
Fetches the datasets contained in the model
}
