% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UD_Server.R
\name{UD_init_state}
\alias{UD_init_state}
\title{Initialize UD Module State}
\usage{
UD_init_state(yaml_file)
}
\arguments{
\item{yaml_file}{App configuration file}
}
\value{
list containing an empty UD state
}
\description{
Creates a list of the initialized module state
}
