% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UD_Server.R
\name{UD_test_mksession}
\alias{UD_test_mksession}
\title{Populate Session Data for Module Testing}
\usage{
UD_test_mksession(session)
}
\arguments{
\item{session}{Shiny session variable (in app) or a list (outside of app)}
}
\value{
The UD portion of the `all_sess_res` returned from \code{\link{ASM_set_app_state}}
}
\description{
Populates the supplied session variable for testing.
}
\examples{
res = UD_test_mksession(session=list())
}
\seealso{
\code{\link{ASM_set_app_state}}
}
